+++ src/application/controllers/users/users.controller.ts
@@ -5,16 +5,13 @@ import {
   getUserById,
   listAllUsers,
   listLastContentsAccessByUserId,
-  listLastSubjectsAccessByUserId,
   updateLastAppAccess,
   updateLastContentAccess,
-  updateLastSubjectAccess,
   updateUser
 } from '../../services/user/user.service';
 import { AppError } from '@/infrastructure/webserver/app-error';
 import { UserInterface } from '@/domain/interfaces/users.interface';
 import { getContentById } from '@/application/services/contents/content.service';
-import { getSubjectById } from '@/application/services/subjects/subject.service';
 
 class UsersController {
   static async getUserById(request: FastifyRequest, reply: FastifyReply) {
@@ -93,19 +90,6 @@ class UsersController {
     }
   }
 
-  static async listLastUserSubjectAccess(
-    request: FastifyRequest,
-    reply: FastifyReply
-  ) {
-    try {
-      const { id: userId } = request.params as { id: string };
-      const subjects = await listLastSubjectsAccessByUserId(userId);
-      return reply.status(200).send(subjects);
-    } catch (error) {
-      throw error;
-    }
-  }
-
   static async updateLastContentAccess(
     request: FastifyRequest,
     reply: FastifyReply
@@ -116,24 +100,6 @@ class UsersController {
       await getUserById(userId);
       await getContentById(contentId);
       await updateLastContentAccess(userId, contentId);
-      return reply
-        .status(200)
-        .send({ message: 'Último acesso atualizado com sucesso' });
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  static async updateLastSubjectAccess(
-    request: FastifyRequest,
-    reply: FastifyReply
-  ) {
-    try {
-      const { id: userId } = request.user;
-      const { subjectId } = request.body as { subjectId: string };
-      await getUserById(userId);
-      await getSubjectById(subjectId);
-      await updateLastSubjectAccess(userId, subjectId);
       return reply
         .status(200)
         .send({ message: 'Último acesso atualizado com sucesso' });
